Move testShardContext from persistence package to history package
